% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_test.R
\name{randomize_test_bias}
\alias{randomize_test_bias}
\title{Randomization Test for Racial Bias}
\usage{
randomize_test_bias(
  race,
  location,
  dist_pop_loc,
  nstops_loc,
  nsim = 1000,
  seed = 5,
  loc,
  prop = TRUE,
  group,
  non = FALSE,
  obs_stops
)
}
\arguments{
\item{race}{vector containing all racial groups;}

\item{location}{vector containing all locations where there are traffic stops;}

\item{nstops_loc}{vector containing number of stops per location;}

\item{nsim}{number of simulation;}

\item{seed}{seed used for simulation for reproducibility;}

\item{loc}{location(s) to use for analysis;}

\item{prop}{Indicate whether the simulation should be shown as proportion or count;}

\item{group}{racial group to conduct test for;}

\item{non}{Used to indicate when interest in say non-white or non-hispanic grouping for test;}

\item{obs_stops}{number/proportion of stops observed in data for the group being tested}

\item{dis_pop_loc}{matrix containing proportion of each race to the DPE per location. race -- rows; location -- columns;}
}
\value{
Simulated p-value and graph showing a randomized distribution of traffic stops
}
\description{
Used to test whether there is any bias in traffic stops of a police officer using randomization test
}
\examples{
race <- c("Asian","Black","Hisp","White","Others");
location <- c("Sunderland","Amherst","Pelham");
dist_pop_loc <- matrix(c(0.075,0.093, 0.128,0.701,0.003,0.075,0.093, 0.128,0.701,0.003,0.075,0.093, 0.128,0.701,0.003), nrow=5, byrow=FALSE);
nstops_loc <- c(50,30,20);
nsim <-1000;
loc <- c("Amherst","Sunderland");
group="Black";
obs_stops <- 8;
randomize_test_bias(race,location,dist_pop_loc,nstops_loc,nsim=1000,seed=5,loc,prop=FALSE,group,non=FALSE,obs_stops);
}
